Improve --help output
The help output grew pretty big and it was hard to find what you
were looking for. By only showing direct subcommands, the help
output at root (secrethub --help) becomes a lot shorter.

I've split the subcommands in management commands and commands
(similarly to the docker --help output), so that it is clear
which commands have another level of subcommands.

Flags that can be used on all commands are no longer displayed in
the help text of the specific commands.

When displaying subcommands, we only show the name and help text.
The usage can be viewed by running --help on the sub-command.