Don't make audit the default command
The default command is executed when no command is supplied
(running secrethub without args in this case). This was used on
the audit command when it was a subcommand of audit. Now that
audit is a root command, we no longer want it to be the default.

Running secrethub without a command prints the help text.