Rework v2 secret templates
This fixes a bug where the closing of a template variable tag
just before the closing of a secret tag returned an error.
Example: {{path/to/secret/${var}}}

The secret templates are now parsed in one go instead of parsing
secret tags first and then parsing the contents of the tag to
look for variables.

The secret template parsing now returns richer error messages, that
include line- and column number where the error occured.

More validation on the contents of the secret- and variable tags
is done while parsing. These errors also include line- and column
number.

dollar signs ($) with a letter, digit or underscore following must
be escaped. This is so that we can add variable support without
brackets later ($var).

$, {, } and \ characters can now be escaped.

Lots of testcases for the parser are added.