Auto detect the template version to use
Instead of defaulting to the latest template version on inject,
detect whether the template contains at least one secret tag in
the old syntax and if so, use the v1 syntax.

To be consistent, the run command now also uses this detection
meganism instead of using a different default based on the env-file
being yml or .env formatted.