Fix races in masked writer
There were two races:
The goroutine of the Run function both receives from  and sends
to the mw.output channel. Altough this works in most cases
because the channel is buffered, this might cause a deadlock,
as the Write goroutine can also send to the channel.
Specifically, when the Write goroutine sends to the channel
after the timeout is triggered in the Run goroutine, but before
Run sends to the channel (when flushing the buffer) a deadlock
occurs.

A second race condition exists in the Flush function of the
writer. However, this could never occur, as Write is not
called after Flush has been called. Therefore, nIn is at its
final state and when nOut is at that same number, it will also
not change any more.
We've fixed it anyway and cleaned up the code a bit.

Co-authored-by: Joris Coenen <jpcoenen@users.noreply.github.com>