Increase sleep time in masked writer test
The current difference of 1ms between the sleep and the timeout
is sometimes not enough to process the write.
The processing of this write takes place in other goroutines and
is not finished when the sleep starts. When processing takes longer
than 1ms, the timeout does not kick in (as expected), because the
timeout is measured after processing of a write.

This should fix undeterministic failure of the test.