Add first version of example-app command
This can be used inside the Getting Started to use as a simple demo app. Can be started with `secrethub example-app serve`. It then serves a web page on the requested ports that tries to connect with httpbin.org. It only succeeds if the APP_USERNAME and APP_PASSWORD environment variables have been set to `demo` and `very-secret-password` respectively.