Create a start repository on signup
This sets up a repository and writes the first secret on signup,
so that the user can start by reading their first secret. This is
intended to smoothen onboarding of new users.

Also, we no longer print the email address should be validated first,
as we have moved this constraint to the moment when users start
collaborating (org invite).