Fix detecting explicitely set values
Kingpin's .Int function returns a pointer to an int. The value is
always set, even when the user does not supply a value. It defaults
to the zero-value: 0.

The new implementation uses a struct that has nil as internal value
when the value is not explicitely set. Therefore, it can return
whether the value was explicitely set with the IsSet function.