Fix run and inject when env contains invalid value
On Windows the environment can contain key-value pair that is not valid
according to the standard. Run and inject used to fail on this because
they try to validate every individual key in the environment. When one
is incorrect, they error. E.g., I encountered the following line in
os.Environ() on Windows: =::=::\

This commits fixes this problem by ignoring any invalid lines in the
environment for injection purposes, but appending them to the
environment of the child process in case of a run.