First test of masker rewrite
TODO:
- Add comments
- Rethink the matching of secrets (flushN() + lookForMatches()). The
current mechanism is pretty inefficient by continuously searching
the part of the buffer that is being flushed + maxLookback number
of bytes before that. This means that if the program that is being
masked uses a lot of small writes, we're doing a regexp lookup
many times over on the same text. The previous implementation did
this better by only checking every byte once. However, this code
was much more complex. There must be a middle-ground here between
complexity and performance.