Fix test failing 50% of times
Success on the executing order of a range over a map. As the order for
looping of maps is undetermined, the test result was also undetermined.

By first performing a sort in the test, the result is no longer
dependent, the assertion no succeeds indepedently of the executing
within this loop.