De-duplicate default envvar to path mapping
Previously, the mapping between envvar keys and paths to where
the values of these envvars should be stored on SecretHub was
defined twice: once for interactive mode and once for non-interactive
mode. This commit refactors the code to use a single definition for
both modes.
This is intended to improve maintainability of the code and to ensure
the default mapping of the interactive mode is the same as the
mapping used in non-interactive mode.