Skip commented lines in dotenv import
Before, one scanner.Scan() was hardcoded to skip the first line
which explained the prompt. Now, the parser instead recognizes
comment lines (starting with #) and skips those.

This has the effect that the comment line can be safely removed
by the user without affecting the result. And also the user can
add more comment lines or comment out lines to ignore them.

One more internal benefit is that we don't have a hardcoded
dependency between the buildFile and buildMap functions sharing
a dependency on how many lines the comment line is.