Include all after "=>" in dotenv import prompt as the secret path
Before, the user could add an extra "=>" in the prompt and everything
following that would be ignored. Now, all after the first "=>" is
used as the path.
The secret read will validate the secret path and throw an error
in case the path contains "=>", as that's not allowed in a secret
path.

So, this commit changes an ignore and continue into an error case.