Use io.Reader and io.Writer in dotenv import helper functions
This prevents loading all contents in memory at the same time.
It removes an unnecessary switch from io.Reader to string and back to
io.Reader.
It gives the helper functions more control, for example, the
buildFile function can now use a tabwriter instead (to be added in
a following commit).