Fix tests using WriteClipboardAutoClear
WriteClipboardAutoClear spawns up a process to clear the clipboard,
using the args of the current process as the base command.
For tests, this caused the test command to be spawned, recursively
spawning more test commands.

We somehow got away with this for a very long while, probably
because the test executes quickly enough to exit before the stack
runs full with processes. However, recently an AUR installation
failed because of this bug in the tests, as AUR installations run
the tests on installation.

This commit fixes the issue by mocking the calls to
WriteClipboardAutoClear in the tests and using a fake implementation
instead, that doesn't spawn new processes.