Refactor: rename clearClipboardAfter and don't pass it to clip writer
It's not just the default value, it's non-configurable for the user.
Because we're using the same value everywhere, there isn't really
a point in having it be configurable.

What's more: we're using the value of the constant in some texts
that are displayed to the user, assuming that this "default" is
the value being used.
I believe it's clearer if we reflect this assumption in the name
of the constant.